backend-2/requirements.txt
fastapi
uvicorn
httpx
transformers
google-generativeai